feat: fix parser.linkLCommandSymbol
